+++ source/filezilla/MainThread.cpp
@@ -143,9 +143,9 @@ BOOL CMainThread::OnThreadMessage(UINT Msg, WPARAM wParam, LPARAM lParam)
     else if (wParam==FZAPI_THREADMSG_PROCESSREPLY)
       m_pControlSocket->ProcessReply();
     else if (wParam==FZAPI_THREADMSG_TRANSFEREND)
-      m_pControlSocket->TransferEnd(lParam);
+      m_pControlSocket->TransferEnd(static_cast<int>(lParam));
     else if (wParam==FZAPI_THREADMSG_CANCEL)
-      m_pControlSocket->Cancel(lParam);
+      m_pControlSocket->Cancel(static_cast<int>(lParam));
     else if (wParam==FZAPI_THREADMSG_DISCONNECT)
       m_pControlSocket->Disconnect();
     else if (wParam==FZAPI_THREADMSG_POSTKEEPALIVE)
